requirements.txt
flask
flask-cors
pymongo[srv]
dnspython
python-dotenv
certifi
